+++ Software/wittypi/init.sh
@@ -11,10 +11,12 @@
 # Description:       This service is used to manage Witty Pi 4 service
 ### END INIT INFO
 
-DAEMON=/home/pi/wittypi/daemon.sh
+# Default paths assume the visIOn layout; install.sh rewrites them to the
+# actual install directory when generating /etc/init.d/wittypi.
+DAEMON=/home/pi/vision/wittypi/daemon.sh
 RETRY_DELAY=15
 MAX_RETRIES=3
-LOG=/home/pi/wittypi/wittyPi.log
+LOG=/home/pi/vision/wittypi/wittyPi.log
 
 # Run the daemon, retrying on non-zero exit. Daemon naturally exits
 # after startup tasks (alarm registers set, schedule revised). A